Replace NonLinearSlider with LogarithmicSlider
