--- docs/src/pages/components/slider/LogarithmicSlider.js
@@ -1,59 +1,0 @@
-import * as React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Typography from '@material-ui/core/Typography';
-import Slider from '@material-ui/core/Slider';
-
-const useStyles = makeStyles({
-  root: {
-    width: 250,
-  },
-});
-
-function getLabel(value) {
-  const units = ['KB', 'MB', 'GB', 'TB'];
-
-  let unitIndex = 0;
-  let scaledValue = value;
-
-  while (scaledValue >= 1024) {
-    unitIndex += 1;
-    scaledValue /= 1024;
-  }
-
-  return `${scaledValue} ${units[unitIndex]}`;
-}
-
-function calculateValue(value) {
-  return 2 ** value;
-}
-
-export default function LogarithmicSlider() {
-  const [value, setValue] = React.useState(10);
-
-  const handleChange = (event, newValue) => {
-    if (typeof newValue === 'number') {
-      setValue(newValue);
-    }
-  };
-
-  const classes = useStyles();
-
-  return (
-    <div className={classes.root}>
-      <Typography id="logarithmic-slider" gutterBottom>
-        Storage: {getLabel(calculateValue(value))}
-      </Typography>
-      <Slider
-        value={value}
-        min={5}
-        step={1}
-        max={30}
-        scale={calculateValue}
-        valueLabelFormat={getLabel}
-        onChange={handleChange}
-        valueLabelDisplay="auto"
-        aria-labelledby="logarithmic-slider"
-      />
-    </div>
-  );
-}
