+++ docs/src/pages/components/slider/NonLinearSlider.tsx
@@ -1,36 +1,57 @@
 import * as React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import Slider from '@material-ui/core/Slider';
 
+const useStyles = makeStyles({
+  root: {
+    width: 250,
+  },
+});
+
 function valueLabelFormat(value: number) {
-  const [coefficient, exponent] = value
-    .toExponential()
-    .split('e')
-    .map((item) => Number(item));
-  return `${Math.round(coefficient)}e^${exponent}`;
+  const units = ['KB', 'MB', 'GB', 'TB'];
+
+  let unitIndex = 0;
+  let scaledValue = value;
+
+  while (scaledValue >= 1024 && unitIndex < units.length - 1) {
+    unitIndex += 1;
+    scaledValue /= 1024;
+  }
+
+  return `${scaledValue} ${units[unitIndex]}`;
+}
+
+function calculateValue(value: number) {
+  return 2 ** value;
 }
 
 export default function NonLinearSlider() {
-  const [value, setValue] = React.useState<number | number[]>(1);
+  const [value, setValue] = React.useState<number>(10);
 
   const handleChange = (
     event: React.SyntheticEvent,
     newValue: number | number[],
   ) => {
-    setValue(newValue);
+    if (typeof newValue === 'number') {
+      setValue(newValue);
+    }
   };
 
+  const classes = useStyles();
+
   return (
-    <div>
+    <div className={classes.root}>
       <Typography id="non-linear-slider" gutterBottom>
-        Temperature range
+        Storage: {valueLabelFormat(calculateValue(value))}
       </Typography>
       <Slider
         value={value}
-        min={0}
-        step={0.1}
-        max={6}
-        scale={(x) => x ** 10}
+        min={5}
+        step={1}
+        max={30}
+        scale={calculateValue}
         getAriaValueText={valueLabelFormat}
         valueLabelFormat={valueLabelFormat}
         onChange={handleChange}
