+++ docs/src/pages/components/slider/slider.md
@@ -93,18 +93,14 @@ The track can be inverted with `track="inverted"`.
 
 {{"demo": "pages/components/slider/TrackInvertedSlider.js"}}
 
-## Non-linear scale
+### Non-linear scale
 
 You can use the `scale` prop to represent the `value` on a different scale.
-For instance, in the following demo, the value _x_ represents the power of _10^x_.
+
+In the following demo, the value _x_ represents the value _2^x_.
+Increasing _x_ by one increases the represented value by factor _2_.
 
 {{"demo": "pages/components/slider/NonLinearSlider.js"}}
-
-### Logarithmic scale
-
-Another use-case for sliders are logarithmic scales. For each step, the value doubles.
-
-{{"demo": "pages/components/slider/LogarithmicSlider.js"}}
 
 ## Accessibility
